src/main/java/code/code250/question228/Solution.java
package code.code250.question228;

import java.util.ArrayList;
import java.util.List;

/**
 * 给定一个无重复元素的有序整数数组，返回数组区间范围的汇总。
 * <p>
 * 示例 1:
 * <p>
 * 输入: [0,1,2,4,5,7]
 * 输出: ["0->2","4->5","7"]
 * 解释: 0,1,2 可组成一个连续的区间; 4,5 可组成一个连续的区间。
 * 示例 2:
 * <p>
 * 输入: [0,2,3,4,6,8,9]
 * 输出: ["0","2->4","6","8->9"]
 * 解释: 2,3,4 可组成一个连续的区间; 8,9 可组成一个连续的区间。
 * <p>
 * 来源：力扣（LeetCode）
 * 链接：https://leetcode-cn.com/problems/summary-ranges
 * 著作权归领扣网络所有。商业转载请联系官方授权，非商业转载请注明出处。
 *
 * @author yan
 * @version 1.0
 * @date 2020/2/12
 */
public class Solution {
    /**
     * 双指针
     *
     * @param nums
     * @return
     */
    public List<String> summaryRanges(int[] nums) {
        List<String> ans = new ArrayList<>();
        int s = 0, e = 0;
        while (e < nums.length) {
            while (e < nums.length - 1 && nums[e] + 1 == nums[e + 1]) {
                e++;
            }
            if (s == e) {
                ans.add(nums[s] + "");
            } else {
                ans.add(nums[s] + "->" + nums[e]);
            }
            e += 1;
            s = e;
        }
        return ans;
    }
}
